wip: start implement with mongodb APM api
Note that this implementation will never works in the current driver
state. Since the driver implement his own queue, the context is lost
between start and end of each operations. The APM api could allows us to match a
specific operation to a span to end it, in theory that would make it
works. However since the original context is lost, downstream span would
never be created since getting the current span after the mongodb one
would get us nothing.